Fix battery reading on statistics

+++ alvr/server/cpp/alvr_server/ClientConnection.cpp
@@ -252,8 +252,8 @@ void ClientConnection::ProcessTimeSync(TimeSync data) {
 				"\"clientFPS\": %.3f, "
 				"\"serverFPS\": %.3f, "
 				"\"batteryHMD\": %llu, "
-				"\"batteryLeft\": %llu, "
-				"\"batteryRight\": %llu"
+				"\"batteryLeft\": %d, "
+				"\"batteryRight\": %d"
 				"} }#\n",
 				m_Statistics->GetPacketsSentTotal(),
 				m_Statistics->GetPacketsSentInSecond(),
@@ -273,8 +273,8 @@ void ClientConnection::ProcessTimeSync(TimeSync data) {
 				m_Statistics->Get(4),  //clientFPS
 				m_Statistics->GetFPS(),
 				m_TrackingInfo.battery,
-				m_TrackingInfo.controller[0].batteryPercentRemaining,
-				m_TrackingInfo.controller[1].batteryPercentRemaining);
+				(int)m_TrackingInfo.controller[0].batteryPercentRemaining,
+				(int)m_TrackingInfo.controller[1].batteryPercentRemaining);
 
 			m_LastStatisticsUpdate = now;
 			m_Statistics->Reset();
